+++ test/dockerValidator.test.ts
@@ -143,6 +143,17 @@ function assertFlagUnknownUnit(diagnostic: Diagnostic, unit: string, duration: s
 	assert.equal(diagnostic.range.end.character, endCharacter);
 }
 
+function assertUnknownHealthcheckFlag(diagnostic: Diagnostic, flag: string, startLine: number, startCharacter: number, endLine: number, endCharacter: number) {
+	assert.equal(diagnostic.code, ValidationCode.UNKNOWN_HEALTHCHECK_FLAG);
+	assert.equal(diagnostic.severity, DiagnosticSeverity.Error);
+	assert.equal(diagnostic.source, source);
+	assert.equal(diagnostic.message, Validator.getDiagnosticMessage_FlagUnknown(flag));
+	assert.equal(diagnostic.range.start.line, startLine);
+	assert.equal(diagnostic.range.start.character, startCharacter);
+	assert.equal(diagnostic.range.end.line, endLine);
+	assert.equal(diagnostic.range.end.character, endCharacter);
+}
+
 function assertInvalidAs(diagnostic: Diagnostic, startLine: number, startCharacter: number, endLine: number, endCharacter: number) {
 	assert.equal(diagnostic.code, ValidationCode.INVALID_AS);
 	assert.equal(diagnostic.severity, DiagnosticSeverity.Error);
@@ -1954,34 +1965,34 @@ describe("Docker Validator Tests", function() {
 				it("unknown flag", function() {
 					let diagnostics = validate("FROM alpine\nHEALTHCHECK --interva=30s CMD ls");
 					assert.equal(diagnostics.length, 1);
-					assertFlagUnknown(diagnostics[0], "interva", 1, 14, 1, 21);
+					assertUnknownHealthcheckFlag(diagnostics[0], "interva", 1, 14, 1, 21);
 
 					// empty value
 					diagnostics = validate("FROM alpine\nHEALTHCHECK --interva= CMD ls");
 					assert.equal(diagnostics.length, 1);
-					assertFlagUnknown(diagnostics[0], "interva", 1, 14, 1, 21);
+					assertUnknownHealthcheckFlag(diagnostics[0], "interva", 1, 14, 1, 21);
 
 					// no equals sign
 					diagnostics = validate("FROM alpine\nHEALTHCHECK --interva CMD ls");
 					assert.equal(diagnostics.length, 1);
-					assertFlagUnknown(diagnostics[0], "interva", 1, 14, 1, 21);
+					assertUnknownHealthcheckFlag(diagnostics[0], "interva", 1, 14, 1, 21);
 
 					// flags are case-sensitive
 					diagnostics = validate("FROM alpine\nHEALTHCHECK --INTERVAL=30s CMD ls");
 					assert.equal(diagnostics.length, 1);
-					assertFlagUnknown(diagnostics[0], "INTERVAL", 1, 14, 1, 22);
+					assertUnknownHealthcheckFlag(diagnostics[0], "INTERVAL", 1, 14, 1, 22);
 
 					diagnostics = validate("FROM alpine\nHEALTHCHECK --RETRIES=3 CMD ls");
 					assert.equal(diagnostics.length, 1);
-					assertFlagUnknown(diagnostics[0], "RETRIES", 1, 14, 1, 21);
+					assertUnknownHealthcheckFlag(diagnostics[0], "RETRIES", 1, 14, 1, 21);
 
 					diagnostics = validate("FROM alpine\nHEALTHCHECK --START-PERIOD=0s CMD ls");
 					assert.equal(diagnostics.length, 1);
-					assertFlagUnknown(diagnostics[0], "START-PERIOD", 1, 14, 1, 26);
+					assertUnknownHealthcheckFlag(diagnostics[0], "START-PERIOD", 1, 14, 1, 26);
 
 					diagnostics = validate("FROM alpine\nHEALTHCHECK --TIMEOUT=30s CMD ls");
 					assert.equal(diagnostics.length, 1);
-					assertFlagUnknown(diagnostics[0], "TIMEOUT", 1, 14, 1, 21);
+					assertUnknownHealthcheckFlag(diagnostics[0], "TIMEOUT", 1, 14, 1, 21);
 				});
 
 				it("flag no value", function() {
@@ -2246,17 +2257,17 @@ describe("Docker Validator Tests", function() {
 					// no equals sign
 					let diagnostics = validate("FROM alpine\nHEALTHCHECK --interva");
 					assert.equal(diagnostics.length, 1);
-					assertFlagUnknown(diagnostics[0], "interva", 1, 14, 1, 21);
+					assertUnknownHealthcheckFlag(diagnostics[0], "interva", 1, 14, 1, 21);
 
 					// empty value
 					diagnostics = validate("FROM alpine\nHEALTHCHECK --interva=");
 					assert.equal(diagnostics.length, 1);
-					assertFlagUnknown(diagnostics[0], "interva", 1, 14, 1, 21);
+					assertUnknownHealthcheckFlag(diagnostics[0], "interva", 1, 14, 1, 21);
 
 					// value specified
 					diagnostics = validate("FROM alpine\nHEALTHCHECK --interva=30s");
 					assert.equal(diagnostics.length, 1);
-					assertFlagUnknown(diagnostics[0], "interva", 1, 14, 1, 21);
+					assertUnknownHealthcheckFlag(diagnostics[0], "interva", 1, 14, 1, 21);
 				});
 			});
 		});